# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CHASSII API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RSVP recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RsvpResult"

  # ── Search & Discovery ────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isFollowingMe:
          type: boolean
        iFollowThem:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        carCount:
          type: integer
        postCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    RsvpResult:
      type: object
      required: [rsvpd]
      properties:
        rsvpd:
          type: boolean
...
    Event:
      type: object
      required: [id, userId, title, type, date, location, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [cars_and_coffee, track_day, meet, group_drive, other]
        date:
          type: string
          format: date-time
        location:
          type: string
        imageUrl:
          type: ["string", "null"]
        rsvpCount:
          type: integer
        hasRsvpd:
          type: boolean
        organizer:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [title, type, date, location]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [cars_and_coffee, track_day, meet, group_drive, other]
        date:
          type: string
          format: date-time
        location:
          type: string
        imageUrl:
          type: string
...